# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel Documentation",
        "description": "Blog API",
        "contact": {
            "name": "Amikel Maxi",
            "email": "maxloversist@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/add/posts": {
            "post": {
                "summary": "Cria um novo post",
                "operationId": "3a61495ed69eaa53e3d4fbb2b7ae3288",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Title is required or Title length must be less than 255 or Content is required ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/edit/posts/{id}": {
            "put": {
                "summary": "Update a post",
                "operationId": "5b4e832af916df26448d4fc6310d1bc3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Post"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Title is required or Content is required or Title must be less than 255 catacteres",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post Id not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "You are not owm of this post or Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/single/posts/{id}": {
            "get": {
                "summary": "Return a specific post",
                "operationId": "3dc5bf818d58e7c32189121189535754",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return the Post",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post id not found"
                    }
                }
            }
        },
...
        "/api/delete/posts/{id}": {
            "delete": {
                "summary": "Delete a post",
                "operationId": "084c46cf24e20dd98465735915487488",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post successfully deleted"
                    },
                    "404": {
                        "description": "Post id not found"
                    },
                    "401": {
                        "description": "You are not owm of this post"
                    }
                }
            }
        },
        "/api/posts/search": {
            "get": {
                "summary": "Return a specific post",
                "operationId": "3879ebb6cd111f8efd219e0fdefc4999",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Like"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Return the list of post based on your query text ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "To search posts, please fill the input box"
                    },
                    "404": {
                        "description": "No post found in this context"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Like": {
                "required": [
                    "like"
                ],
                "properties": {
                    "id": {
                        "description": "The like id",
                        "type": "integer"
                    },
                    "like": {
                        "description": "Like (true or false)",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "Post": {
                "required": [
                    "title",
                    "content",
                    "user_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the post",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "description": "Title of the post",
                        "type": "string",
                        "example": "My post"
                    },
                    "content": {
                        "description": "Content of the post",
                        "type": "text",
                        "example": "Description of rst postmy fi"
                    },
                    "user_id": {
                        "description": "The user ID, owm of the post",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        }
    }
}